capitals/requirements.txt
flask
flask-cors
oauth2client==2.0.0
google-cloud
